Rename gspconsumer to pvliveconsumer for clarity
