+++ .all-contributorsrc
@@ -34,7 +34,7 @@
     }
   ],
   "contributorsPerLine": 7,
-  "projectName": "GSPConsumer",
+  "projectName": "GSPConsumer", 
   "projectOwner": "openclimatefix",
   "repoType": "github",
   "repoHost": "https://github.com",
