+++ .bumpversion.cfg
@@ -4,6 +4,6 @@ tag = True
 current_version = 1.2.1
 message = Bump version: {current_version} → {new_version} [skip ci]
 
-[bumpversion:file:gspconsumer/__init__.py]
+[bumpversion:file:pvliveconsumer/__init__.py]
 search = __version__ = "{current_version}"
 replace = __version__ = "{new_version}"
