+++ .github/workflows/test-docker.yaml
@@ -18,7 +18,7 @@ jobs:
         run: docker compose -f test-docker-compose.yml build
 
       - name: Run tests inside the container
-        run: docker compose -f test-docker-compose.yml run gspconsumer
+        run: docker compose -f test-docker-compose.yml run pvliveconsumer
         env: # Or as an environment variable
           API_KEY: ${{ secrets.API_KEY }}
           SYSTEM_ID: ${{ secrets.SYSTEM_ID }}
