+++ setup.py
@@ -1,4 +1,4 @@
-"""Setup file for gspconsumer package."""
+"""Setup file for pvliveconsumer package."""
 from pathlib import Path
 
 from setuptools import find_packages, setup
@@ -7,14 +7,14 @@ this_directory = Path(__file__).parent
 install_requires = (this_directory / "requirements.txt").read_text().splitlines()
 
 # get version
-with open("gspconsumer/__init__.py") as f:
+with open("pvliveconsumer/__init__.py") as f:
     for line in f:
         if line.startswith("__version__"):
             _, _, version = line.replace("'", "").split()
             version = version.replace('"', "")
 
 setup(
-    name="gspconsumer",
+    name="pvliveconsumer",
     version=version,
     packages=find_packages(),
     install_requires=install_requires,
