+++ tests/test_nighttime.py
@@ -1,4 +1,4 @@
-from gspconsumer.nitghtime import make_night_time_zeros
+from pvliveconsumer.nitghtime import make_night_time_zeros
 from datetime import datetime
 from nowcasting_datamodel.models.gsp import LocationSQL, GSPYieldSQL
 import pandas as pd
